refactor(proxy): split http bridge mixin slices
